+++ coreapi/sal/sal.c
@@ -896,7 +896,7 @@ static int line_get_value(const char *input, const char *key, char *value, size_
 	char *equal;
 	size_t len;
 	if (!end) len=strlen(input);
-	else len=end +1 -input;
+	else len=(size_t)(end + 1 - input);
 	*read=len;
 	strncpy(line,input,MIN(len,sizeof(line)));
 	equal=strchr(line,'=');
